Alterado de usuarioService para authService.register no método de registro

+++ app/src/main/java/com/ecommerce/app/service/address/AddressService.java
@@ -26,7 +26,7 @@ public class AddressService {
                 .collect(Collectors.toList());
     }
 
-    public AddressDTO getAddressById(Long id){
+    public AddressDTO getAddressById(Integer id){
 //        Optional<Address> address = addressRepository.findById(id);
 //        return address.map(this::convertToDTO).orElse(null);
     	Address address = addressRepository.findById(id)
@@ -52,7 +52,7 @@ public class AddressService {
         return convertToDTO(address);
     }
 
-    public AddressDTO updateAddress(Long id, AddressDTO addressDTO){
+    public AddressDTO updateAddress(Integer id, AddressDTO addressDTO){
         Optional<Address> addressOptional = addressRepository.findById(id);
         if(addressOptional.isPresent()){
             Address address = addressOptional.get();
@@ -75,7 +75,7 @@ public class AddressService {
         return null;
     }
 
-    public void deleteAddress(Long id){
+    public void deleteAddress(Integer id){
         addressRepository.deleteById(id);
     }
